fix: remove headers merging (#16)
It is not allowed to merge multiple HTTP headers into a single
HAR::Header with values joined with ", ". The count of objects in
`headers` field should be exactly the same as the count of headers in
the original HTTP request or response.

Merging headers can create a malformed "Set-Cookie" header with multiple
cookies that can't be parsed by Crystal's HTTP::Cookies class and can
even raise an exception.